backend/requirements.txt
# Core FastAPI & Server
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Database & ORM
sqlalchemy[asyncio]==2.0.31
asyncpg==0.29.0 # Replace if using different DB driver
alembic==1.13.2

# Data Validation & Settings
pydantic[email]==2.8.2
pydantic-settings==2.3.4

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Add other runtime dependencies below (with specific versions)
# Example: emails==0.6